+++ tests/testthat/test-dependencies.R
@@ -1,0 +1,3 @@
+test_that("reactable_extras_dependency should return the correct dependency", {
+  expect_snapshot(reactable_extras_dependency())
+})