+++ NEWS.md
@@ -1,1 +1,11 @@
 # reactable.extras (development version)
+
+# reactable.extras 0.1.0.9000
+
+## Server-Side Processing
+
+* `reactable_extras_ui()` and `reactable_extras_server()` now allows for server-side processing of data.
+
+## Miscellaneous
+
+* Added a `NEWS.md` file to track changes to the package.